pyproject.toml
[project]
name = "rag22"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "open-clip-torch>=3.3.0",
    "pypdf>=6.7.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
]
